docs/casos-uso/requirements.txt
# Dependencias para convertir casos de uso a formato Word
python-docx==0.8.11
